+++ WPNest/WPNest/Services/ITimestampProvider.cs
@@ -1,6 +1,6 @@
 ﻿namespace WPNest.Services {
 
 	public interface ITimestampProvider {
-		long GetTimestamp();
+		double GetTimestamp();
 	}
 }
